all tasks with functional programming
